+++ include/engine/Geometry.h
@@ -11,53 +11,53 @@ class Size;
 
 class Point : public std::enable_shared_from_this<Point> {
 public:
-    double x;
-    double y;
+    long double x;
+    long double y;
 
     Point();
-    Point(double dim);
-    Point(double x, double y);
+    Point(long double dim);
+    Point(long double x, long double y);
     Point(const Size& size);
 
     // In Radians
-    static std::shared_ptr<Point> createFromAngle(double angle);
+    static std::shared_ptr<Point> createFromAngle(long double angle);
     
     Point& operator+=(const Point& right);
     Point& operator-=(const Point& right);
     Point operator+(const Point& right) const;
     Point operator-(const Point& right) const;
     Point operator-() const;
-    Point operator*(double a) const;
-    Point operator/(double a) const;
+    Point operator*(long double a) const;
+    Point operator/(long double a) const;
     bool operator==(const Point& right) const;
 
     operator std::string() const;
     operator SDL_FPoint() const;
 
-    void setPoint(double x, double y);
+    void setPoint(long double x, long double y);
     bool equals(const Point& target) const;
     
-    bool fuzzyEquals(const Point& target, double variance) const;
+    bool fuzzyEquals(const Point& target, long double variance) const;
 
-    double getLength() const;
+    long double getLength() const;
 
-    double getDistance(const Point& other) const;
+    long double getDistance(const Point& other) const;
 
-    double getAngle() const;
+    long double getAngle() const;
 
     Point normalise() const;
 
-    Point rotateByAngle(double angle);
+    Point rotateByAngle(long double angle);
 };
 
 class Size : public std::enable_shared_from_this<Size> {
 public:
-    double width;
-    double height;
+    long double width;
+    long double height;
 
     Size();
-    Size(double dim);
-    Size(double width, double height);
+    Size(long double dim);
+    Size(long double width, long double height);
     Size(const Point& point);
 
     Size& operator+=(const Size& right);
@@ -65,14 +65,14 @@ public:
     Size operator+(const Size& right);
     Size operator-(const Size& right);
     Size operator-() const;
-    Size operator*(double a) const;
-    Size operator/(double a) const;
+    Size operator*(long double a) const;
+    Size operator/(long double a) const;
     bool operator==(const Size& right) const;
 
     operator std::string() const;
     operator SDL_FPoint() const;
 
-    void setSize(double width, double height);
+    void setSize(long double width, long double height);
     bool equals(const Size& target) const;
 };
 
@@ -82,7 +82,7 @@ public:
     Size size;
 
     Rect();
-    Rect(double x, double y, double width, double height);
+    Rect(long double x, long double y, long double width, long double height);
     Rect(const Point& origin, const Size& size);
 
     bool operator==(const Rect& right) const;
@@ -91,22 +91,22 @@ public:
 
     operator std::string() const;
 
-    void setRect(double x, double y, double width, double height);
+    void setRect(long double x, long double y, long double width, long double height);
     void setOrigin(const Point& point);
     void setSize(const Size& size);
 
     bool equals(const Rect& other) const;
-    double getMinX() const;
-    double getMaxX() const;
-    double getMinY() const;
-    double getMaxY() const;
-    std::tuple<double, double, double, double> getBounds() const;
+    long double getMinX() const;
+    long double getMaxX() const;
+    long double getMinY() const;
+    long double getMaxY() const;
+    std::tuple<long double, long double, long double, long double> getBounds() const;
 
     bool containsPoint(const Point& point) const;
     bool intersectsRect(const Rect& rect) const;
 
-    double getX() const;
-    double getY() const;
-    double getWidth() const;
-    double getHeight() const;
+    long double getX() const;
+    long double getY() const;
+    long double getWidth() const;
+    long double getHeight() const;
 };
