+++ include/engine/Node.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <vector>
 
 #include "Object.h"
@@ -12,19 +13,19 @@ public:
     static std::shared_ptr<Node> create();
 
 
-    virtual void setZOrder(long long zOrder);
-    virtual long long getZOrder();
+    virtual void setZOrder(int64_t zOrder);
+    virtual int64_t getZOrder();
 
-    virtual void setScale(double scale);
-    virtual double getScale();
+    virtual void setScale(long double scale);
+    virtual long double getScale();
 
     virtual void setPosition(Point position);
     virtual Point getPosition() const;
-    virtual void setPosition(double x, double y);
-    virtual void setPositionX(double x);
-    virtual double getPositionX() const;
-    virtual void setPositionY(double y);
-    virtual double getPositionY() const;
+    virtual void setPosition(long double x, long double y);
+    virtual void setPositionX(long double x);
+    virtual long double getPositionX() const;
+    virtual void setPositionY(long double y);
+    virtual long double getPositionY() const;
 
     virtual void setAnchorPoint(Point anchorPoint);
     virtual Point getAnchorPoint() const;
@@ -35,40 +36,40 @@ public:
     virtual void setVisible(bool visible);
     virtual bool isVisible() const;
 
-    virtual void setRotation(double rotation);
-    virtual double getRotation() const;
+    virtual void setRotation(long double rotation);
+    virtual long double getRotation() const;
     
 
     virtual void addChild(std::shared_ptr<Node> child);
-    virtual void addChild(std::shared_ptr<Node> child, long long zOrder);
-    virtual void addChild(std::shared_ptr<Node> child, long long zOrder, long long tag);
+    virtual void addChild(std::shared_ptr<Node> child, int64_t zOrder);
+    virtual void addChild(std::shared_ptr<Node> child, int64_t zOrder, int64_t tag);
     // Returns nullptr if child does not exist
-    virtual std::shared_ptr<Node> getChildByTag(long long tag);
+    virtual std::shared_ptr<Node> getChildByTag(int64_t tag);
     virtual std::vector<std::shared_ptr<Node>> getChildren();
     
-    virtual unsigned long long getChildrenCount() const;
+    virtual uint64_t getChildrenCount() const;
     
     virtual void setParent(std::shared_ptr<Node> parent);
     virtual std::shared_ptr<Node> getParent();
     
     virtual void removeFromParent();
     virtual void removeChild(std::shared_ptr<Node> child);
-    virtual void removeChildByTag(long long tag);
+    virtual void removeChildByTag(int64_t tag);
     virtual void removeAllChildren();
     
-    virtual void reorderChild(std::shared_ptr<Node> child, long long zOrder);
+    virtual void reorderChild(std::shared_ptr<Node> child, int64_t zOrder);
     
     virtual void sortAllChildren();
     
     
-    virtual void setTag(long long tag);
+    virtual void setTag(int64_t tag);
     
     virtual std::shared_ptr<void> getUserData();
     virtual void setUserData(std::shared_ptr<void> userData);
 
 
     virtual void cleanup();
-    virtual void draw(const double dt);
+    virtual void draw(const long double dt);
 
     
     Rect getRect();
@@ -76,18 +77,18 @@ public:
 protected:
     Node();
 
-    double m_rotation;
-    double m_scale;
+    long double m_rotation;
+    long double m_scale;
     Point m_position;
     Point m_anchorPoint;
     Size m_contentSize;
     
-    long long m_zOrder;
+    int64_t m_zOrder;
     
     std::vector<std::shared_ptr<Node>> m_children;
     std::shared_ptr<Node> m_parent;
     
-    long long m_tag;
+    int64_t m_tag;
     std::shared_ptr<void> m_userData;
 
     bool m_visible;
